Updated to move unit tests to core module as this is the logic being tested
