+++ worker-markup-core/src/test/java/com/github/cafdataprocessing/worker/markup/core/ConversationIndexParserTest.java
@@ -13,10 +13,8 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.hpe.caf.worker.markup;
+package com.github.cafdataprocessing.worker.markup.core;
 
-import com.github.cafdataprocessing.worker.markup.core.ConversationIndexParser;
-import com.github.cafdataprocessing.worker.markup.core.ConversationIndexParserException;
 import org.apache.commons.codec.binary.Base64;
 import org.junit.Test;
