+++ worker-markup-core/src/test/java/com/github/cafdataprocessing/worker/markup/core/EmailSplitterTest.java
@@ -13,10 +13,8 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.hpe.caf.worker.markup;
+package com.github.cafdataprocessing.worker.markup.core;
 
-import com.github.cafdataprocessing.worker.markup.core.EmailSplitter;
-import com.github.cafdataprocessing.worker.markup.core.JepExecutor;
 import org.jdom2.Document;
 import org.jdom2.JDOMException;
 import org.jdom2.input.SAXBuilder;
