+++ worker-markup-core/src/test/java/com/github/cafdataprocessing/worker/markup/core/TestUtility.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.hpe.caf.worker.markup;
+package com.github.cafdataprocessing.worker.markup.core;
 
 import java.io.File;
 import java.io.IOException;
