+++ worker-markup-core/src/test/java/com/github/cafdataprocessing/worker/markup/core/XmlParsingHelperTest.java
@@ -13,10 +13,8 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.hpe.caf.worker.markup;
+package com.github.cafdataprocessing.worker.markup.core;
 
-import com.github.cafdataprocessing.worker.markup.core.MarkupHeadersAndBody;
-import com.github.cafdataprocessing.worker.markup.core.XmlParsingHelper;
 import static org.junit.Assert.assertEquals;
 
 import org.junit.Test;
