+++ worker-markup-core/src/test/resources/xml/CondensedHeaderMultiLangMapping.xml
