+++ worker-markup-core/src/test/resources/xml/MultipleEmailTags.xml
