+++ worker-markup-core/src/test/resources/xml/OtherLanguageHeaderSurroundedWithAsterisks.xml
