+++ worker-markup-core/src/test/resources/xml/OtherLanguageHeaderSurroundedWithAsterisksExpected.xml
