+++ worker-markup-core/src/test/resources/xml/ThreadedEmailWithIllegalChars.xml
