+++ worker-markup-core/src/test/resources/xml/ThreadedEmailWithIllegalCharsMarkedup.xml
