+++ worker-markup-core/src/test/resources/xml/headerAsterisk.xml
