+++ worker-markup-core/src/test/resources/xml/headerAsteriskExpected.xml
