+++ worker-markup-core/src/test/resources/xml/headerAsteriskFromValueOnTwoLines.xml
