+++ worker-markup-core/src/test/resources/xml/headerAsteriskFromValueOnTwoLinesExpected.xml
