+++ worker-markup-core/src/test/resources/xml/headerAsteriskFromValueOnTwoLinesFromNotFirst.xml
